@@ -14,5 +14,5 @@ flutter:
     pluginClass: FlutterStatusbarcolorPlugin
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=0.1.4 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=2.0.6 <3.0.0"
